Se modificó la forma en que se leen los pulsos. Ahora el contador LPTIM4 encargado de contarlos empieza a correr desde el inicio del programa y no se detiene nunca. Lo que se toma ahora como nuevos pulsos a agregar es la diferencia entre los pulsos que se tenía antes de irse a dormir y los que se tiene al despertarse y realizar todas las operaciones que debe hacer antes de volver a dormirse. A niveles prácticos servirá para que siempre se estén leyendo el timer en busca de pulsos y por lo tanto se pueda medir la velocidad en versiones futuras del firmware. Versión 1.2.9

+++ libs/fm_factory/fm_factory.c
@@ -708,7 +708,12 @@ void fm_factory_modify_k_lin_subs(sel_digit_k_lin_t digit_k_lin, sel_k k_sel)
     }
 }
 
-void fm_factory_modify_pulse_acm_ttl(int added_pulses)
+/*
+ * @brief Actualizo el valor de los pulsos de volumen.
+ * @param Pulsos a añadir a los ya existentes.
+ * @retval None
+ */
+void fm_factory_modify_pulse_acm_ttl(uint16_t added_pulses)
 {
     acm_config.pulse += added_pulses;
     ttl_config.pulse += added_pulses;
